requirements.txt
flask
beautifulsoup4
openai
google-generativeai
pillow
requests
tqdm
python-dotenv
qdrant-client
python-dotenv